c-rewrite: implement initial control flow
This commit contains the skeleton of high-level control flow, mostly
taken from the Rust version. Arguments are parsed, and, if help doesn't
need to be given, all the specified input files are opened and
'assembled' (though no assembly is present in this commit). Errors are
saved in a tagged union and logged to stderr.