store enums and flags in enums instead of macros
In C, enums begin at 0 and increment upwards. Stylistic name changes
aside, the behavior is identical to the #define blocks, but the compiler
is better able to type-check our work.

This commit also apparently adds in a newline at the end of register.h,
which should be the case in every file so that the number of newlines
directly corresponds with the number of lines in general.