c-rewrite: begin parser (implement hlt and nop)
This commit begins a skeleton of parser architecture. Assembly is
specifically line-based and buffered as such.

The error RE_NOINS error is introduced, and indicates that the paser
encountered an instruction it doesn't recognize. Unlike RE_FNOOPEN, it
stores an *owned copy* of a string (which indicates what the
unrecognized instruction is), and must be freed after use (thus some
extra logic in err_free_asc (free associated data).

Both the hlt and nop instructions are able to be parsed with this commit.