c: fix invalid free by reading only whole lines
While the previous BSD-specific fgetln implementation returned memory
managed by the FILE itself, getline returned a dynamically allocated
buffer that must be manually freed. Moving the pointer instruction-by-
instruction lead to a pointer within the allocation, not to the
beginning, causing problems with an invalid free.

The lazy solution (used here) is to only expect one instruction on each
line. Instructions delimited by semicolons are a Problem For Later.