c: allow characters to follow register
This limitation is defined every time a register is parsed, so, for
instance, a comma may follow the first register in an rrmovq, but not
the last register. If a disallowed character is present, a BADREG error
is recorded.

Note that allowed characters are not *necessarily* present, meaning that
statements like `rrmovq %rbp %rsp` *are* parsed in a forgiving manner.