c: refactor register reading destination into parameter
When reading a register into a gen_ins, it may be read into the upper or
lower half of the register byte. If read most-significant before
least-significant, this can be accomplished agnostically via bit
shifting in the caller. If however, least significant must be read
*before* most significant (as will be required by mrmovq), read_reg must
itself know whether it's reading into the upper or lower half-byte.