write tests and package for nix
This commit introduces a tests/ directory with a test case, and some
scripts in a nix flake to run them with `nix build`.

If nix isn't used, the tests may be run by placing the compiled binary
on PATH, cd'ing to the tests subdirectory, and executing runnner.sh.
(Nix packaging just exists for CI convenience.)