asm: parse immediate values as signed
Essentially, we want to allow negative immediate values for irmovq's.
(This commit modifies the irmovq test to verify that this works as
expected.)

Note that because jump targets share the same immediate value parsing
logic, this commit also makes those signed, despite negative values not
making sense there. While integer range is definitely not an issue given
realistic input sizes, an error is introduced if the jump target is
negative.