asm/style: extract function read_dest
This allows all instructions that read a jump destination (jXX and call)
to share some more code, and allow for an easier time writing the label
isolation instructions.

Also included in this commit is some misc improvements to error handling
that i just cannot be bothered to split out. (Use a 'template' parameter
to record file info.)